All forecast tests pass except test_numpy; no Prophet
