+++ flaml/automl/__init__.py
@@ -1,4 +1,5 @@
-from .automl import AutoMLState, SearchState, logger_formatter, size
+from .automl import logger_formatter, size
+from .state import SearchState, AutoMLState
 from ._factory import AutoML
 
-__all__ = ["AutoML", "AutoMLState", "SearchState", "logger_formatter", "size"]
+__all__ = ["AutoML", "logger_formatter", "size"]
