+++ flaml/automl/_factory.py
@@ -1,17 +1,17 @@
 from . import tasks
 from .automl import AutoML as AutoMLGeneric
 from .time_series import AutoMLTS
-from ..data import TS_FORECASTREGRESSION
+from ..data import TS_FORECAST
 
 
 # Hackity hack hack
-# The below is ultra-grim, but allows us to break the TS_FORECASTREGRESSION
+# The below is ultra-grim, but allows us to break the TS_FORECAST
 # logic into its own class which maintaining backwards compatibility of
 # the public interface to AutoML
 class AutoML(AutoMLGeneric):
     def __init__(self, **settings):
         super().__init__(**settings)
-        if self._settings["task"] in TS_FORECASTREGRESSION:
+        if self._settings["task"] in TS_FORECAST:
             AutoMLTS.__init__(self, **settings)
 
     def fit(self, *args, **kwargs):
@@ -22,9 +22,9 @@ class AutoML(AutoMLGeneric):
             task_name = args[5]
         # Is it in self._settings?
         task_name = task_name or self._settings["task"]
-        task = tasks.task_factory(task_name)
+        self.task = tasks.task_factory(task_name)
 
-        self.__class__ = task.AUTOML_CLASS
+        self.__class__ = self.task.AUTOML_CLASS
         self.fit(*args, **kwargs)
