+++ flaml/automl/tasks.py
@@ -1,13 +1,13 @@
 from .automl import AutoML as AutoMLGeneric
 from .time_series import AutoMLTS
-from ..data import TS_FORECASTREGRESSION
+from ..data import TS_FORECAST
 
 
 class Task:
     AUTOML_CLASS = AutoMLGeneric
 
     def __init__(self, task_name):
-        self.task_name = task_name
+        self.name = task_name
 
 
 class TaskTS(Task):
@@ -15,6 +15,6 @@ class TaskTS(Task):
 
 
 def task_factory(task_name: str) -> Task:
-    if task_name in TS_FORECASTREGRESSION:
+    if task_name in TS_FORECAST:
         return TaskTS(task_name)
     return Task(task_name)
